Feature/docker compose (#8)
* feat: deploy

+++ backend/src/controllers/role.controller.ts
@@ -1,11 +1,14 @@
 import { NextFunction, Request, Response } from 'express';
+
 import { RoleReq, UpdateRoleReq } from 'interfaces/role.interface';
 import RoleService from '../services/role.service';
-import { ResponseMap, HttpCodeSuccess } from '../utils/const';
+import { ResponseMap } from '../utils/const';
+import { Administrator } from '../utils/index';
 import { pageCompute } from '../utils/pageCompute';
 import ResourceService from '../services/resource.service';
+import ResponseHandler from '../utils/responseHandler';
 
-const { Success, ParamsError, RoleExisted } = ResponseMap
+const { ParamsError, RoleExisted, SystemEmptyError } = ResponseMap
 
 class RoleController {
   public RoleService = new RoleService();
@@ -14,34 +17,29 @@ class RoleController {
   public getRoles = async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { namespace, role, name, resource, current, pageSize } = req.query as any;
+
+      // 检查必填参数
       if (!namespace) {
-        res.status(HttpCodeSuccess).json(ParamsError);
-        return;
+        return ResponseHandler.error(res, ParamsError);
       }
-      const params: any = { 
+
+      // 计算分页参数
+      const params: any = {
         namespace,
-        ...pageCompute(current, pageSize) 
-      }
-      if (role) {
-        params.role = role
-      }
-      if (name) {
-        params.name = name
-      }
-      if (resource) {
-        params.resource = resource.split(",")
-      }
+        ...pageCompute(current, pageSize),
+      };
+
+      // 可选参数添加到 params 中
+      if (role) params.role = role;
+      if (name) params.name = name;
+      if (resource) params.resource = resource.split(",");
+
       const getData: any = await this.RoleService.findWithAllChildren(params);
       const { rows, count } = getData;
       // 聚合资源id
       const resourceIds = rows.map((item: any) => item.dataValues.resources).flat();
       if(resourceIds.length === 0){
-        res.status(HttpCodeSuccess).json({ 
-          ...Success, 
-          data: rows || [],
-          total: 0 
-        });
-        return 
+        return ResponseHandler.success(res, { data: rows||[], total: count});
       }
       // 根据resource资源列表
       const resourceList: any = await this.ResourceService.findAllByIds(resourceIds);
@@ -51,11 +49,7 @@ class RoleController {
       rows.forEach((item: any) => {
         delete item.dataValues.resources
       });
-      res.status(HttpCodeSuccess).json({ 
-        ...Success, 
-        data: rows || [],
-        total: count 
-      });
+      return ResponseHandler.success(res, { data: rows||[], total: count } );
 		} catch (error) {
 			next(error);
 		}
@@ -63,21 +57,18 @@ class RoleController {
 
   public createRole = async (req: Request, res: Response, next: NextFunction) => {
     try {
-      const query: RoleReq = req.body;
+      const params: RoleReq = req.body;
       const remoteUser = req.headers['remoteUser'] as string;
-      const params  = query as RoleReq;
       const { namespace, role, name } = params;
       
       if (!params || !namespace || !role) {
-        res.status(HttpCodeSuccess).json(ParamsError);
-        return;
+        return ResponseHandler.error(res, ParamsError);
       }
 
       // 查询当前roleName是否存在
       const checkRole = await this.RoleService.findRoleByRoleName({namespace, roleName: role, name})
       if(checkRole){
-        res.status(HttpCodeSuccess).json(RoleExisted);
-        return;
+        return ResponseHandler.error(res, RoleExisted);
       }else{
         const createData: any = await this.RoleService.create(
           {
@@ -85,10 +76,7 @@ class RoleController {
             operator: remoteUser
           }
         );
-        res.status(HttpCodeSuccess).json({ 
-          ...Success, 
-          data: createData
-        });
+        return ResponseHandler.success(res, createData);
       }
 		} catch (error) {
 			next(error);
@@ -99,25 +87,27 @@ class RoleController {
     try {
       const body: UpdateRoleReq = req.body;
       const remoteUser = req.headers['remoteUser']
-      const { id, namespace, role, permissions, name, describe } = body;
+      const { id, namespace, role: newRole, permissions, name, describe } = body;
+      
       if(!id || !namespace){
-        return res.status(HttpCodeSuccess).json({
-          ...ParamsError
-        });
+        return ResponseHandler.error(res, ParamsError);
       }
+      // 查询如果是超级管理员，则不允许更改role
+      const role = await this.RoleService.findRoleById({namespace, id})
+      if(!role){
+        return ResponseHandler.error(res, SystemEmptyError, 'Role not found');
+      }
+      const replaceRole = role.dataValues.role === Administrator ? Administrator : newRole;
       const response: any = await this.RoleService.update({
         id,
         namespace,
-        role,
+        role: replaceRole,
         permissions,
         name,
         describe,
         operator: remoteUser
       })
-      res.status(HttpCodeSuccess).json({ 
-        ...Success, 
-        data: response
-      });
+      return ResponseHandler.success(res, response);
 		} catch (error) {
 			next(error);
 		}
@@ -126,14 +116,23 @@ class RoleController {
   public deleteRole = async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { id, namespace } = req.body;
+
+      // 检查是否为管理员，管理员不允许删除
+      const role = await this.RoleService.findRoleById({namespace, id})
+      
+      if(!role){
+        return ResponseHandler.error(res, SystemEmptyError, 'Role not found');
+      }
+      
+      if(role.dataValues.role === 'admin'){
+        return ResponseHandler.error(res, SystemEmptyError, 'admin not allow to delete');
+      }
+
       const getData: any = await this.RoleService.deleteSelf({
         id,
 				namespace
       });
-      res.status(HttpCodeSuccess).json({ 
-        ...Success, 
-        data: getData
-      });
+      return ResponseHandler.success(res, getData);
 		} catch (error) {
 			next(error);
 		}
@@ -146,8 +145,7 @@ class RoleController {
       const params  = query as any;
       
       if (!params) {
-        res.status(HttpCodeSuccess).json(ParamsError);
-        return;
+        return ResponseHandler.error(res, ParamsError);
       }
 
 			const createData: any = await this.RoleService.createRolePermission(
@@ -156,10 +154,7 @@ class RoleController {
           operator: remoteUser
         }
 			);
-      res.status(HttpCodeSuccess).json({ 
-        ...Success, 
-        data: createData
-      });
+      return ResponseHandler.success(res, createData);
 		} catch (error) {
 			next(error);
 		}
